Many things :     - Move test of file system in test.c     - Improve makefiles, now we have target for test and fs.img is not always recreated     - In consequences, change test in fs_create     - Some clean and avoiding gcc warnings
On branch master

Changes to be committed:
	modified:   .gitignore
	modified:   common/makefile
	modified:   kernel/fs_api.c
	modified:   kernel/kernel.c
	modified:   kernel/makefile
	modified:   kernel/test.c
	modified:   kernel/test.h
	modified:   makefile
	deleted:    tools/.gitignore
	modified:   tools/add_files.sh
	deleted:    tools/del.sh
	modified:   tools/fs_add.c
	modified:   tools/fs_create.c
	modified:   tools/makefile
	modified:   tools/scenario.sh
	deleted:    tools/txt_files1/a.txt
	deleted:    tools/txt_files1/hello.txt
	deleted:    tools/txt_files1/raed.txt
	deleted:    tools/txt_files1/s.txt

+++ kernel/fs_api.c
@@ -214,7 +214,7 @@ int file_seek(int fd, uint offset) {
 			printf("The file descriptor in unknown (file_seek)\n");
 			return -1;
 		}
-		if (offset > file_descriptor[fd].file_size) {
+		if (offset > (uint) file_descriptor[fd].file_size) {
 			offset = file_descriptor[fd].file_size;
 		}
